Dynamic Changes + Project Restructuring
- In this I have made the following dynamically change as your temperature or pressure change
  - Temperature Background Color
  - Weather Background Color
  - Weather Icon
  - Weather Icon Color
- In doing so, I also was able to combine all 3 weather checker functions into 1. This new function, titled baro_dynamic_decision, sets all of the weather colors and icon in one place. This reduces complexity.
- Additionally, I separated the two pages away from the main function and renamed the second class to be "WeatherPage". So now there are three .dart files, grouping the separate parts together.
